+++ euporie/core/utils.py
@@ -5,11 +5,16 @@ from __future__ import annotations
 from itertools import chain
 from typing import TYPE_CHECKING, Sequence, TypeVar, overload
 
+from prompt_toolkit.mouse_events import MouseButton, MouseEventType
 from upath import UPath
 
 if TYPE_CHECKING:
     from os import PathLike
-    from typing import Iterable, Optional, Union
+    from typing import Callable, Iterable, Optional, Union
+
+    from prompt_toolkit.key_binding.key_bindings import NotImplementedOrNone
+    from prompt_toolkit.layout.mouse_handlers import MouseHandler
+    from prompt_toolkit.mouse_events import MouseEvent
 
 T = TypeVar("T")
 
@@ -57,3 +62,17 @@ def parse_path(path: "Optional[Union[str, PathLike]]") -> "Optional[UPath]":
     except AttributeError:
         pass
     return upath
+
+
+def on_click(func: "Callable") -> "MouseHandler":
+    """Return a mouse handler which call a given function on click."""
+
+    def _mouse_handler(mouse_event: "MouseEvent") -> "NotImplementedOrNone":
+        if (
+            mouse_event.button == MouseButton.LEFT
+            and mouse_event.event_type == MouseEventType.MOUSE_UP
+        ):
+            return func()
+        return NotImplemented
+
+    return _mouse_handler